Reformat consumable and enum data
